Move `merkle` types into their correct headers.
